outsource general methods
